# LB8/lab8/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info": {
    "title": "LeraX",
    "version": "1.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "tags" : [ {
    "name" : "Lera"
  } ],
  "paths" : {
    "/api/Lera" : {
      "get" : {
        "tags" : [ "Lera" ],
        "summary" : "Get All Users",
        "description" : "create 24.07.2020",
        "responses" : {
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/inline_response_200"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Lera" ],
        "summary" : "Update User by id",
        "description" : "create 24.07.2020",
        "requestBody" : {
          "content" : {
            "application/json-patch+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "User updated",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            }
          },
          "400" : {
            "description" : "User not added"
          }
        }
      },
      "post" : {
        "tags" : [ "Lera" ],
        "summary" : "Add User",
        "description" : "create 24.07.2020",
        "requestBody" : {
          "content" : {
            "application/json-patch+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "User updated",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            }
          },
          "400" : {
            "description" : "User not added"
          }
        }
      }
    },
    "/api/Lera/{id}" : {
      "get" : {
        "tags" : [ "Lera" ],
        "summary" : "Get User by id",
        "description" : "create 24.07.2020",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The User id",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/User/properties/id"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "User found",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            }
          },
          "404" : {
            "description" : "User not found"
          }
        }
      },
      "delete" : {
        "tags" : [ "Lera" ],
        "summary" : "Delete User by id",
        "description" : "create 24.07.2020",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "The User id",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "$ref" : "#/components/schemas/User/properties/id"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "User deleted",
            "content" : {
              "text/plain" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            }
          },
          "404" : {
            "description" : "User not found"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "User" : {
        "required" : [ "email", "id", "password", "role", "status" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "description" : "Users Id",
            "format" : "int32",
            "example" : 101
          },
          "lastName" : {
            "type" : "string",
            "description" : "Users Surname",
            "nullable" : true,
            "example" : "Adam Smith"
          },
          "firstName" : {
            "type" : "string",
            "description" : "Users Name",
            "example" : "Adam"
          },
          "email" : {
            "type" : "string",
            "description" : "Users e-mail",
            "format" : "email",
            "example" : "xxx@yyy.com"
          },
          "password" : {
            "type" : "string",
            "description" : "password",
            "example" : "p@$$w0rd"
          },
          "status" : {
            "type" : "string",
            "description" : "active/passive",
            "example" : "active"
          },
          "role" : {
            "type" : "string",
            "description" : "admin/customer/HR",
            "example" : "customer"
          }
        },
        "example" : {
          "lastName" : "Adam Smith",
          "firstName" : "Adam",
          "password" : "p@$$w0rd",
          "role" : "customer",
          "id" : 101,
          "email" : "xxx@yyy.com",
          "status" : "active"
        }
      },
      "Userslist" : {
        "type" : "array",
        "nullable" : true,
        "items" : {
          "$ref" : "#/components/schemas/User"
        }
      },
      "inline_response_200" : {
        "type" : "object",
        "properties" : {
          "userslist" : {
            "$ref" : "#/components/schemas/Userslist"
          }
        },
        "example" : {
          "userslist" : [ {
            "lastName" : "Adam Smith",
            "firstName" : "Adam",
            "password" : "p@$$w0rd",
            "role" : "customer",
            "id" : 101,
            "email" : "xxx@yyy.com",
            "status" : "active"
          }, {
            "lastName" : "Adam Smith",
            "firstName" : "Adam",
            "password" : "p@$$w0rd",
            "role" : "customer",
            "id" : 101,
            "email" : "xxx@yyy.com",
            "status" : "active"
          } ]
        }
      }
    }
  }
}